requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
WTForms==3.0.1
psycopg[binary]==3.2.10
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.0.0
gunicorn==20.1.0
paystack==1.5.0
requests==2.31.0